+++ lib/src/fields/smart_text_field.dart
@@ -17,6 +17,7 @@ class SmartTextField extends StatefulWidget {
     this.validators = const [],
     this.asyncValidators = const [],
     this.asyncValidationDebounce,
+    this.autovalidateMode = AutovalidateMode.onUnfocus,
     this.errorAnimation,
     this.enabled = true,
     this.decoration = const InputDecoration(),
@@ -44,6 +45,7 @@ class SmartTextField extends StatefulWidget {
   final List<SmartValidator<String>> validators;
   final List<SmartAsyncValidator<String>> asyncValidators;
   final Duration? asyncValidationDebounce;
+  final AutovalidateMode autovalidateMode;
   final SmartErrorAnimation? errorAnimation;
   final bool enabled;
   final InputDecoration decoration;
@@ -132,6 +134,7 @@ class _SmartTextFieldState extends State<SmartTextField> {
       validators: widget.validators,
       asyncValidators: widget.asyncValidators,
       asyncValidationDebounce: widget.asyncValidationDebounce,
+      autovalidateMode: widget.autovalidateMode,
       errorAnimation: widget.errorAnimation,
       enabled: widget.enabled,
       focusNode: widget.focusNode,